cached-sts-token gets command to retrieve AWS keys
When the cache of credential has expired, we need to call
`generate-sts-token.sh` again, but with the root credentials
that would assume the role.

Sometimes such command is expensive or tedious (e.g. federated login)
and we want to run it only when the cache is expired.

We pass this ROOT_AWS_CREDENTIALS_COMMAND, that would print
the root AWS that would be `eval` before calling
`generate-sts-token.sh`. Usually it would be the `pass keys/creds.sh`
command.